添加任务修改逻辑，解决相关bug

+++ renrenentity/src/main/java/com/renrentui/renrenentity/RenRenTask.java
@@ -28,7 +28,7 @@ public class RenRenTask {
 
     private Date endTime;
 
-    private Double taskCycle;
+    private Integer taskCycle;
 
     private Integer availableCount;
 
@@ -39,20 +39,19 @@ public class RenRenTask {
     private Integer taskTotalCount;
 
     private Long snapshotTemplateId;
-
     private String link;
 
     private Short paymentMethod;
-    private int auditCycle;
+    private Integer auditCycle;
 
     private String taskNote;
     private Long targetPeople;
     private String companySummary;
-    public int getAuditCycle() {
+    public Integer getAuditCycle() {
 		return auditCycle;
 	}
 
-	public void setAuditCycle(int auditCycle) {
+	public void setAuditCycle(Integer auditCycle) {
 		this.auditCycle = auditCycle;
 	}
 
@@ -190,11 +189,11 @@ public class RenRenTask {
 		this.endTime = endTime;
 	}
 
-	public Double getTaskCycle() {
+	public Integer getTaskCycle() {
 		return taskCycle;
 	}
 
-	public void setTaskCycle(Double taskCycle) {
+	public void setTaskCycle(Integer taskCycle) {
 		this.taskCycle = taskCycle;
 	}
 
@@ -237,14 +236,18 @@ public class RenRenTask {
 	public void setTaskTotalCount(Integer taskTotalCount) {
 		this.taskTotalCount = taskTotalCount;
 	}
-
+	/**
+	 * 快照模板id
+	 * @param templateId
+	 */
 	public Long getSnapshotTemplateId() {
 		return snapshotTemplateId;
 	}
-
+	/**
+	 * 快照模板id
+	 * @param templateId
+	 */
 	public void setSnapshotTemplateId(Long snapshotTemplateId) {
 		this.snapshotTemplateId = snapshotTemplateId;
-	}
-
-    
+	} 
 }
